+++ src/onegov.gazette/HISTORY.rst
@@ -28,6 +28,9 @@ Changelog
 - Sanitizes HTML much stricter.
   [msom]
 
+- Allows to delete users with official notices.
+  [msom]
+
 0.0.4 (2017-08-03)
 ~~~~~~~~~~~~~~~~~~~